% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm2sls.R
\name{fit2sls}
\alias{fit2sls}
\title{2SLS Regression}
\usage{
fit2sls(y, X, Z, wt = NULL, singular.ok = FALSE, qr = TRUE)
}
\arguments{
\item{y}{The response variable, a vector.}

\item{X}{The explanatory-variable model matrix.}

\item{Z}{The instrumental-variable model matrix,}

\item{wt}{An optional weight vector.}

\item{singular.ok}{Is perfect collinearity tolerable in either stage regression?}

\item{qr}{Include the QR decomposition in the returned object.}
}
\value{
A list with the following components:
\describe{
  \item{\code{n}}{number of cases}
  \item{\code{p}}{number of columns of \code{X}}
  \item{\code{q}}{number of columns of \code{Z}}
  \item{\code{qr}}{QR decomposition for stage-2 regression}
  \item{\code{rank}}{rank of model matrix for stage-2 regression}
  \item{\code{qr.1}}{QR decomposition for stage-1 regression}
  \item{\code{rank.1}}{rank of \code{Z}}
  \item{\code{coefficients}}{estimated regression coefficients}
  \item{\code{coefficients.1}}{coefficient matrix from stage11 regression}
  \item{\code{vcov}}{covariance matrix of estimated coefficients}
  \item{\code{df.residual}}{residual degrees of freedom for stage-2 regression}
  \item{\code{sigma}}{residual standard deviation}
  \item{\code{residuals}}{response residuals}
  \item{\code{fitted}}{model fitted values}
  \item{\code{residuals.1}}{matrix of stage-1 residuals}
  \item{\code{fitted.1}}{matrix of stage=1 fitted values}
  \item{\code{residuals.2}}{stage-2 residuals}
  \item{\code{fitted.2}}{stage-2 fitted values}
}
}
\description{
The work-horse 2SLS function, not normally meant to be called directly
but rather via \code{\link{lm2sls}}.
}
\examples{
m <- with(Kmenta, fit2sls(Q, cbind(1, P, D), cbind(1, D, F, A)))
names(m)
m$coefficients # estimates
sqrt(diag(m$vcov)) # standard errors
}
\seealso{
\code{\link{lm2sls}}
}
\author{
John Fox \email{jfox@mcmaster.ca}
}
