% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm2sls.R
\name{Rsq}
\alias{Rsq}
\title{R-Squares Analog}
\usage{
Rsq(model, ...)
}
\arguments{
\item{model}{A model object.}

\item{...}{Possible arguments for specific methods.}
}
\description{
R-Squares Analog
}
\examples{
Rsq(lm2sls(Q ~ P + D, ~ D + F + A, data=Kmenta))
}
