% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.2sls.R
\name{influence.2sls}
\alias{influence.2sls}
\alias{rstudent.2sls}
\alias{cooks.distance.2sls}
\alias{dfbeta.influence.2sls}
\alias{dfbeta.2sls}
\alias{hatvalues.2sls}
\alias{rstudent.influence.2sls}
\alias{hatvalues.influence.2sls}
\alias{cooks.distance.influence.2sls}
\title{Deletion Diagnostic Methods for \code{"2sls"} Objects}
\usage{
\method{influence}{2sls}(model, sigma. = n <= 1000, type = c("stage2",
  "both"), ...)

\method{rstudent}{2sls}(model, ...)

\method{cooks.distance}{2sls}(model, ...)

\method{dfbeta}{influence.2sls}(model, ...)

\method{dfbeta}{2sls}(model, ...)

\method{hatvalues}{2sls}(model, type = c("stage2", "both"), ...)

\method{rstudent}{influence.2sls}(model, ...)

\method{hatvalues}{influence.2sls}(model, ...)

\method{cooks.distance}{influence.2sls}(model, ...)
}
\arguments{
\item{model}{A \code{"2sls"} or \code{"influence.2sls"} object.}

\item{sigma.}{A if \code{TRUE} (the default for 1000 or fewer cases), the deleted value
of the residual standard deviation is computed for each case; if \code{FALSE}, the
overall residual standard deviation is used to compute other deletion diagnostics.}

\item{type}{If \code{"stage2"} (the default) hatvalues are for the second stage regression;
if \code{"both"} the hatvalues are the geometric mean of the casewise hatvalues for the
two stages.}

\item{...}{Not used.}
}
\value{
In the case of \code{influence.2sls}, an object of class \code{"influence.2sls"}
with the following components:
\describe{
\item{\code{dfbeta}}{influence on coefficients.}
\item{\code{sigma}}{deleted values of residual standard deviation.}
\item{\code{dffits}}{overall influence on regression coefficients.}
\item{\code{cookd}}{Cook's distances.}
\item{\code{hatvalues}}{hatvalues.}
\item{\code{rstudent}}{Studentized residuals.}
}
In the case of other methods, such as \code{rstudent.2sls} or
\code{rstudent.influence.2sls}, the corresponding diagnostic statistics.
}
\description{
Methods for computing deletion diagnostics for 2SLS regression.
It's generally more efficient to compute the diagnostics via the \code{influence}
method and then to extract the various specific diagnostics with the methods for
\code{"influence.2sls"} objects.
}
